+++ internal/core/editor/editor.go
