+++ internal/core/editor/editor_test.go
