+++ internal/core/exit/cmd.go
