+++ internal/core/git/stash.go
