+++ internal/core/git/stash_test.go
