+++ internal/core/git/status_test.go
