+++ internal/core/textwrap/builder.go
