+++ internal/core/textwrap/builder_test.go
