+++ internal/core/textwrap/characterwrap.go
