+++ internal/core/textwrap/characterwrap_test.go
