+++ internal/core/textwrap/renderer.go
