+++ internal/core/textwrap/wordwrap.go
