+++ internal/core/textwrap/wordwrap_test.go
