+++ internal/core/textwrap/wrapper.go
