+++ internal/core/ui/components/container/content.go
