+++ internal/core/ui/components/dialog/cmd.go
