+++ internal/core/ui/components/dialog/confirm/confirm.go
@@ -4,8 +4,8 @@ import (
 	"github.com/charmbracelet/bubbles/key"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/michaelhass/gitglance/internal/ui/dialog"
-	styles "github.com/michaelhass/gitglance/internal/ui/style"
+	"github.com/michaelhass/gitglance/internal/core/ui/components/dialog"
+	styles "github.com/michaelhass/gitglance/internal/core/ui/style"
 )
 
 const (
