+++ internal/core/ui/components/dialog/confirm/key.go
