+++ internal/core/ui/components/dialog/content.go
