+++ internal/core/ui/components/list/container.go
@@ -1,17 +1,22 @@
-package diff
+package list
 
 import (
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/michaelhass/gitglance/internal/ui/container"
+	"github.com/michaelhass/gitglance/internal/core/ui/components/container"
 )
 
-// ContainerContent is a wrapper to use the commit ui as container.ContainerContent.
+// ContainerContent is a wrapper to use the filelist ui as container.ContainerContent.
 type ContainerContent struct {
 	Model
 }
 
-func NewContent(model Model) ContainerContent {
+func NewContainerContent(model Model) ContainerContent {
 	return ContainerContent{Model: model}
+}
+
+func NewContainer(model Model) container.Model {
+	containerContent := NewContainerContent(model)
+	return container.New(containerContent)
 }
 
 func (c ContainerContent) Update(msg tea.Msg) (container.Content, tea.Cmd) {
