+++ internal/core/ui/components/list/msg.go
