+++ internal/domain/diff/key.go
